added new parent class FigureParent.java, and solution also added GetArea.java interface, FigureParents implements Figure and GetArea interfaces, and other class extends FigureParent.java
