--- src/main/java/core/basesyntax/AbstractFigure.java
@@ -1,13 +1,0 @@
-package core.basesyntax;
-
-public abstract class AbstractFigure implements Figure {
-    private final String color;
-
-    public AbstractFigure(String color) {
-        this.color = color;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
