+++ src/main/java/core/basesyntax/Figure.java
@@ -2,10 +2,4 @@ package core.basesyntax;
 
 public interface Figure {
     void draw();
-
-    double findArea();
-
-    String getColor();
-
-    double roundTo(double value);
 }
