+++ src/main/java/core/basesyntax/GetArea.java
@@ -1,0 +1,7 @@
+package core.basesyntax;
+
+public interface GetArea {
+    double findArea();
+
+    double roundTo(double value);
+}
